requirements.txt
streamlit>=1.30.0
setuptools>=68.0.0
wheel>=0.40.0
numpy>=1.26.0
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
openai==0.28.0
pandas>=2.1.0
requests>=2.31.0
scipy>=1.12.0
matplotlib>=3.8.0
pandas>=2.1.0
tqdm>=4.66.0
